// webfact/src/app/titulos/consultas/models/titulos-consulta.ts
export interface TitulosConsultaModel {
    tipoConsulta: string;
    periodoInicial: Date;
    peridoFinal: Date;
    cadastroInicial: string;
    cadastroFinal: string;
    tipoOperacao: number;
    tipoTitulo: string;
    valor: number;
    cnpj: string;
    cedente: string;
    titulo: string;
}